+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Add DateIntervalUtils to stringify a DateInterval, and DateIntervalFactory for shorthand creation
 * Add DateTimeImmutableFactory::zeroMicros() to create / modify a DT with microseconds truncated to zero
 * Add DateTimeImmutableFactory::fromIso() to strictly parse ISO 8601 / RFC 3339 date-time strings
 * Fix bug where devices with long user agents generated a database exception when attempting to create a session.